Finished html of contact page
I finished the html of the contact page. I also added basic css to the page to figure out the layout and design of the page. I fixed some problems such as page overflow and padding issues. I want to maybe add an image slider at the end of the page in the contact social section.